refactor device digital twins
