+++ src/app/devices/pnp/components/deviceDigitalTwin/digitalTwinDetail.scss
@@ -2,8 +2,8 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License
  **********************************************************/
- @import '../../../css/themes';
- @import '../../../css/variables';
+ @import '../../../../css/themes';
+ @import '../../../../css/variables';
 
 .digitaltwin-pivot {
    height: 100%;
