+++ src/app/devices/pnp/components/deviceDigitalTwin/digitalTwinDetail.spec.tsx
@@ -6,9 +6,9 @@ import 'jest';
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import { DigitalTwinDetail } from './digitalTwinDetail';
-import * as pnpStateContext from '../../../shared/contexts/pnpStateContext';
-import { PnpStateInterface, pnpStateInitial } from '../state';
-import { SynchronizationStatus } from '../../../api/models/synchronizationStatus';
+import * as pnpStateContext from '../../../../shared/contexts/pnpStateContext';
+import { PnpStateInterface, pnpStateInitial } from '../../state';
+import { SynchronizationStatus } from '../../../../api/models/synchronizationStatus';
 
 const search = '?id=device1&componentName=foo&interfaceId=urn:iotInterfaces:com:interface1;1';
 const pathname = `/#/devices/deviceDetail/ioTPlugAndPlay/${search}`;
