+++ src/app/devices/pnp/components/pnp.tsx
@@ -8,7 +8,7 @@ import { ROUTE_PARTS } from '../../../constants/routes';
 import { getDeviceIdFromQueryString, getInterfaceIdFromQueryString, getComponentNameFromQueryString, getModuleIdentityIdFromQueryString } from '../../../shared/utils/queryStringHelper';
 import { getModelDefinitionAction } from '../actions';
 import { PnpStateContextProvider } from '../../../shared/contexts/pnpStateContext';
-import { DigitalTwinDetail } from './digitalTwinDetail';
+import { DigitalTwinDetail } from './deviceDigitalTwin/digitalTwinDetail';
 import { useAsyncSagaReducer } from '../../../shared/hooks/useAsyncSagaReducer';
 import { pnpReducer } from '../reducer';
 import { pnpSaga } from '../saga';
@@ -16,7 +16,7 @@ import { pnpStateInitial } from '../state';
 import { RepositoryLocationSettings } from '../../../shared/global/state';
 import { useGlobalStateContext } from '../../../shared/contexts/globalStateContext';
 import { getRepositoryLocationSettings } from '../../../modelRepository/dataHelper';
-import { DigitalTwinInterfacesList } from './digitalTwinInterfacesList';
+import { DigitalTwinInterfacesList } from './deviceDigitalTwin/digitalTwinInterfacesList';
 import { BreadcrumbRoute } from '../../../navigation/components/breadcrumbRoute';
 import '../../../css/_digitalTwinInterfaces.scss';
 import { dispatchGetTwinAction } from '../utils';
